chore: move settings to navigation bar
